Make Switchboard an embeddable WSGI application.
